Fix Windows printf format specifier for ssize_t
Changed %zd to %d with explicit cast to (int) for bytes_read variable.
Windows MSVC doesn't properly support %zd format specifier.

This resolves the build error:
warning C4477: 'printf' : format string '%zd' requires an argument
of type 'unsigned __int64', but variadic argument 1 has type 'ssize_t'

+++ tests/mock_restreamer.c
@@ -159,11 +159,11 @@ static void *server_thread(void *arg) {
 
     if (bytes_read > 0) {
       buffer[bytes_read] = '\0';
-      printf("[MOCK] Received %zd bytes, handling request\n", bytes_read);
+      printf("[MOCK] Received %d bytes, handling request\n", (int)bytes_read);
       handle_request(client_fd, buffer);
       printf("[MOCK] Response sent\n");
     } else {
-      printf("[MOCK] No data received (bytes_read=%zd)\n", bytes_read);
+      printf("[MOCK] No data received (bytes_read=%d)\n", (int)bytes_read);
     }
 
 #ifdef _WIN32
